requirements.txt
flask~=3.1.2
openai
pyyaml~=6.0.3
flask-cors~=6.0.2
tiktoken~=0.12.0
livereload~=2.7.1
gunicorn